INGM-679 remove duplicated fixtures
